requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
WTForms==3.2.1
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.2.1
pycountry==24.6.1
qrcode==8.2
pillow==12.0.0
PyMySQL==1.1.2
cryptography==46.0.3
